Fixing url issue
