+++ examples/accounts/index.js
@@ -2,7 +2,7 @@ const nrg = require('../..')
 
 const app = nrg.createApp({
   name: 'Accounts Example',
-  baseUrl: 'http://localhost:9999/',
+  port: 9999,
   log: { level: 'error' },
   db: {
     connection: {
