+++ lib/app.js
@@ -14,14 +14,14 @@ const { Print } = require('@ianwalter/print')
 const { oneLine } = require('common-tags')
 const enableDestroy = require('server-destroy')
 const { serveStatic, serveWebpack } = require('./middleware/client')
-const defaultConfig = require('./defaultConfig')
+const config = require('./config')
 
 // If running in a CLI context, parse the JSON string to get the CLI options.
 const nrgCli = process.env.NRG_CLI && JSON.parse(process.env.NRG_CLI)
 
 module.exports = function createApp (options = {}) {
   // Combine defaults, CLI-supplied options, and user-supplied options.
-  const cfg = merge({}, defaultConfig, options, nrgCli)
+  const cfg = merge({}, config(options), options, nrgCli)
 
   // Create the Koa app instance.
   const app = new Koa()
@@ -177,13 +177,11 @@ module.exports = function createApp (options = {}) {
   // Add a start method to the app that makes it easy to start listening for
   // connections.
   app.start = function start () {
-    const { name, port, host } = app.context.cfg
-
     // TODO: comment
     app.server = http.createServer(app.callback())
 
     return new Promise(resolve => {
-      app.server.listen(port, host, err => {
+      app.server.listen(app.context.cfg.port, app.context.cfg.host, err => {
         if (err) {
           app.logger.error(err)
           process.exit(1)
@@ -191,17 +189,21 @@ module.exports = function createApp (options = {}) {
 
         // Update the port in the config in case it wasn't specified and Node
         // has used a random port.
-        const { port } = app.server.address()
-        app.context.cfg.port = port
+        if (!app.context.cfg.port) {
+          const { port } = app.server.address()
+          app.context.cfg.port = port
+        }
 
         // Set the server URL (the local URL which can be different from the
         // base URL) so that whatever is starting the server (e.g. tests) can
         // easily know what URL to use.
-        app.server.url = `http://${host}:${port}`
+        app.server.url = app.context.cfg.hostUrl
 
         // NOTE: pino is not including the URL in the message if passed as the
         // second param here.
-        app.logger.info(`${name} now serving requests at: ${app.server.url}`)
+        app.logger.info(oneLine`
+          ${app.context.cfg.name} now serving requests at: ${app.server.url}
+        `)
 
         // Add a destroy method to the server instance.
         // https://github.com/nodejs/node/issues/2642
