+++ lib/middleware/emailVerification.js
@@ -1,16 +1,13 @@
-const querystring = require('querystring')
 const merge = require('@ianwalter/merge')
+const createUrl = require('@ianwalter/url')
 const { ValidationError } = require('../errors')
 const { generateToken, insertToken } = require('./token')
 const { sendEmail } = require('./email')
 
 function handleEmailVerificationEmail (ctx, next, options) {
-  const route = ctx.cfg.url.origin + options.path
-  const query = querystring.stringify({
-    email: ctx.data.email,
-    token: ctx.result.token
-  })
-  const link = `${route}?${query}`
+  const url = createUrl(ctx.cfg.baseUrl, options.path)
+  url.search = { email: ctx.data.email, token: ctx.result.token }
+  const link = url.href
   const { passwordReset } = ctx.cfg.email.templates
   const { name } = ctx.result
   const body = merge({}, passwordReset, { name, action: { button: { link } } })
