+++ lib/middleware/forgotPassword.js
@@ -1,15 +1,12 @@
-const querystring = require('querystring')
 const merge = require('@ianwalter/merge')
+const createUrl = require('@ianwalter/url')
 
 function handlePasswordResetEmail (ctx, next, options) {
-  const route = ctx.cfg.url.origin + options.path
-  const query = querystring.stringify({
-    email: ctx.data.email,
-    token: ctx.result.token
-  })
-  const link = `${route}?${query}`
   const { passwordReset } = ctx.cfg.email.templates
   const { name } = ctx.result
+  const url = createUrl(ctx.cfg.baseUrl, options.path)
+  url.search = { email: ctx.data.email, token: ctx.result.token }
+  const link = url.href
   const body = merge({}, passwordReset, { name, action: { button: { link } } })
   const html = ctx.mailgen.generate({ body })
   const subject = options.subject || `${ctx.cfg.name} Password Reset`
