+++ lib/middleware/httpsRedirect.js
@@ -1,9 +1,12 @@
+const createUrl = require('@ianwalter/url')
+
 function httpsRedirect (ctx, next) {
   const protocol = ctx.get('x-forwarded-proto')
   if (protocol && protocol.toLowerCase() === 'http') {
     ctx.log.debug('Redirecting http request to https')
     ctx.status = 301
-    return ctx.redirect(ctx.cfg.url.origin + ctx.url)
+    const url = createUrl(ctx.cfg.baseUrl, ctx.url)
+    return ctx.redirect(url.href)
   }
   return next()
 }
