+++ lib/middleware/ssr.js
@@ -5,6 +5,7 @@ const requireFromString = require('require-from-string')
 const devalue = require('devalue')
 const pkgDir = require('pkg-dir')
 const cloneable = require('@ianwalter/cloneable')
+const createUrl = require('@ianwalter/url')
 
 const appModule = module.parent.parent.parent || module.parent.parent
 const mainDir = pkgDir.sync(appModule.filename)
@@ -40,7 +41,7 @@ function serveSsr (options) {
 
     // Add the "full" URL on the context so that a router can use it with the
     // URL class.
-    ctx.fullUrl = ctx.cfg.url.origin + ctx.url
+    ctx.fullUrl = createUrl(ctx.cfg.baseUrl, ctx.url)
 
     // TODO: comment
     const result = await options.renderer(ctx, next)
